update to npm packages format
