+++ sdk/README.md
@@ -2,6 +2,10 @@
 
 ## Installation
 
+yarn add @proofofpassport/sdk
+
+## Development
+
 ```bash
 yarn install-sdk
 ```
