+++ sdk/index.ts
@@ -1,0 +1,9 @@
+import { ProofOfPassportWeb2Verifier } from './ProofOfPassportWeb2Verifier';
+import { ProofOfPassportWeb2Inputs } from './ProofOfPassportWeb2Verifier';
+import { ProofOfPassportWeb3Verifier } from './ProofOfPassportWeb3Verifier';
+
+export {
+    ProofOfPassportWeb2Verifier,
+    ProofOfPassportWeb3Verifier,
+    ProofOfPassportWeb2Inputs
+};
